+++ execution/executors/executor_test/executor_test.go
@@ -545,8 +545,9 @@ func TestSimpleInsertAndLimitExecutionMultiTable(t *testing.T) {
 }
 
 func TestHashTableIndex(t *testing.T) {
+	common.TempSuppressOnMemStorage = true
+
 	diskManager := disk.NewDiskManagerTest()
-	defer diskManager.ShutDown()
 	log_mgr := recovery.NewLogManager(&diskManager)
 	bpm := buffer.NewBufferPoolManager(uint32(32), diskManager, log_mgr) //, recovery.NewLogManager(diskManager), access.NewLockManager(access.REGULAR, access.PREVENTION))
 
@@ -678,6 +679,8 @@ func TestHashTableIndex(t *testing.T) {
 		})
 	}
 
+	diskManager.ShutDown()
+	common.TempSuppressOnMemStorage = false
 }
 
 func TestSimpleDelete(t *testing.T) {
@@ -1095,7 +1098,6 @@ func TestInsertUpdateMix(t *testing.T) {
 }
 
 func TestAbortWIthDeleteUpdate(t *testing.T) {
-	os.Stdout.Sync()
 	diskManager := disk.NewDiskManagerTest()
 	defer diskManager.ShutDown()
 	log_mgr := recovery.NewLogManager(&diskManager)
@@ -1237,7 +1239,6 @@ func TestAbortWIthDeleteUpdate(t *testing.T) {
 }
 
 func TestSimpleHashJoin(t *testing.T) {
-	os.Stdout.Sync()
 	diskManager := disk.NewDiskManagerTest()
 	defer diskManager.ShutDown()
 	log_mgr := recovery.NewLogManager(&diskManager)
@@ -1565,8 +1566,10 @@ func timeoutPanic() {
 }
 
 func TestConcurrentTransactionExecution(t *testing.T) {
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 
 	shi := samehada.NewSamehadaInstanceForTesting()
 	shi.GetLogManager().ActivateLogging()
@@ -1656,8 +1659,10 @@ func TestConcurrentTransactionExecution(t *testing.T) {
 }
 
 func TestTestTableGenerator(t *testing.T) {
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 
 	shi := samehada.NewSamehadaInstanceForTesting()
 	shi.GetLogManager().ActivateLogging()
@@ -1694,8 +1699,10 @@ func TestTestTableGenerator(t *testing.T) {
 
 func TestSimpleAggregation(t *testing.T) {
 	// SELECT COUNT(colA), SUM(colA), min(colA), max(colA) from test_1;
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 
 	shi := samehada.NewSamehadaInstanceForTesting()
 	shi.GetLogManager().ActivateLogging()
@@ -1764,8 +1771,10 @@ func TestSimpleAggregation(t *testing.T) {
 
 func TestSimpleGroupByAggregation(t *testing.T) {
 	// SELECT count(colA), colB, sum(C) FROM test_1 Group By colB HAVING count(colA) > 100
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 
 	shi := samehada.NewSamehadaInstanceForTesting()
 	shi.GetLogManager().ActivateLogging()
@@ -1843,8 +1852,10 @@ func TestSimpleGroupByAggregation(t *testing.T) {
 
 func TestSeqScanWithMultiItemPredicate(t *testing.T) {
 	// SELECT colA, colB colC FROM test_1 WHERE (colA > 500 AND colB < 5) OR (NOT colC >= 1000)
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 
 	shi := samehada.NewSamehadaInstanceForTesting()
 	shi.GetLogManager().ActivateLogging()
@@ -1999,8 +2010,10 @@ func TestInsertAndSpecifiedColumnUpdate(t *testing.T) {
 }
 
 func TestInsertAndSpecifiedColumnUpdatePageMoveCase(t *testing.T) {
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 
 	shi := samehada.NewSamehadaInstanceForTesting()
 	shi.GetLogManager().ActivateLogging()
@@ -2169,8 +2182,10 @@ func TestInsertAndSpecifiedColumnUpdatePageMoveCase(t *testing.T) {
 
 func TestSimpleSeqScanAndOrderBy(t *testing.T) {
 	// SELECT a, b, FROM test_1 ORDER BY a, b
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 
 	shi := samehada.NewSamehadaInstanceForTesting()
 	shi.GetLogManager().ActivateLogging()
@@ -2276,8 +2291,10 @@ func TestSimpleSeqScanAndOrderBy(t *testing.T) {
 }
 
 func TestSimpleSetNullToVarchar(t *testing.T) {
-	os.Remove("test.db")
-	os.Remove("test.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("test.db")
+		os.Remove("test.log")
+	}
 
 	shi := samehada.NewSamehadaInstanceForTesting()
 	shi.GetLogManager().ActivateLogging()
