+++ samehada/samehada_test/samehada_test.go
@@ -2,6 +2,7 @@ package samehada_test
 
 import (
 	"fmt"
+	"github.com/ryogrid/SamehadaDB/common"
 	"github.com/ryogrid/SamehadaDB/samehada"
 	testingpkg "github.com/ryogrid/SamehadaDB/testing"
 	"os"
@@ -10,9 +11,13 @@ import (
 
 // TODO: (SDB) need to check query result (TestInsertAndMultiItemPredicateSelect)
 func TestInsertAndMultiItemPredicateSelect(t *testing.T) {
+	common.TempSuppressOnMemStorage = true
+
 	// clear all state of DB
-	os.Remove("example.db")
-	os.Remove("example.log")
+	if !common.EnableOnMemStorage || common.TempSuppressOnMemStorage == true {
+		os.Remove("example.db")
+		os.Remove("example.log")
+	}
 
 	db := samehada.NewSamehadaDB("example", 200)
 	db.ExecuteSQLRetValues("CREATE TABLE name_age_list(name VARCHAR(256), age INT);")
@@ -33,13 +38,16 @@ func TestInsertAndMultiItemPredicateSelect(t *testing.T) {
 	samehada.PrintExecuteResults(results5)
 
 	db.Shutdown()
+	common.TempSuppressOnMemStorage = false
 }
 
 // TODO: (SDB) need to check query result (TestHasJoinSelect)
 func TestHasJoinSelect(t *testing.T) {
 	// clear all state of DB
-	os.Remove("example.db")
-	os.Remove("example.log")
+	if !common.EnableOnMemStorage {
+		os.Remove("example.db")
+		os.Remove("example.log")
+	}
 
 	db := samehada.NewSamehadaDB("example", 200)
 	db.ExecuteSQLRetValues("CREATE TABLE id_name_list(id INT, name VARCHAR(256));")
@@ -67,9 +75,12 @@ func TestHasJoinSelect(t *testing.T) {
 }
 
 func TestSimpleDelete(t *testing.T) {
+	common.TempSuppressOnMemStorage = true
 	// clear all state of DB
-	os.Remove("example.db")
-	os.Remove("example.log")
+	if !common.EnableOnMemStorage || common.TempSuppressOnMemStorage == true {
+		os.Remove("example.db")
+		os.Remove("example.log")
+	}
 
 	db := samehada.NewSamehadaDB("example", 200)
 	db.ExecuteSQLRetValues("CREATE TABLE name_age_list(name VARCHAR(256), age INT);")
@@ -87,9 +98,12 @@ func TestSimpleDelete(t *testing.T) {
 }
 
 func TestSimpleUpdate(t *testing.T) {
+	common.TempSuppressOnMemStorage = true
 	// clear all state of DB
-	os.Remove("example.db")
-	os.Remove("example.log")
+	if !common.EnableOnMemStorage || common.TempSuppressOnMemStorage == true {
+		os.Remove("example.db")
+		os.Remove("example.log")
+	}
 
 	db := samehada.NewSamehadaDB("example", 200)
 	db.ExecuteSQLRetValues("CREATE TABLE name_age_list(name VARCHAR(256), age INT);")
@@ -105,12 +119,17 @@ func TestSimpleUpdate(t *testing.T) {
 	testingpkg.SimpleAssert(t, len(results1) == 3)
 
 	db.Shutdown()
+	common.TempSuppressOnMemStorage = false
 }
 
 func TestRebootWithLoadAndRecovery(t *testing.T) {
+	common.TempSuppressOnMemStorage = true
+
 	// clear all state of DB
-	os.Remove("/tmp/todo.db")
-	os.Remove("/tmp/todo.log")
+	if !common.EnableOnMemStorage || common.TempSuppressOnMemStorage == true {
+		os.Remove("/tmp/todo.db")
+		os.Remove("/tmp/todo.log")
+	}
 
 	db := samehada.NewSamehadaDB("/tmp/todo", 200)
 	db.ExecuteSQLRetValues("CREATE TABLE name_age_list(name VARCHAR(256), age INT);")
@@ -151,12 +170,18 @@ func TestRebootWithLoadAndRecovery(t *testing.T) {
 
 	// close db and log file
 	db3.Shutdown()
+
+	common.TempSuppressOnMemStorage = false
 }
 
 func TestRebootAndReturnIFValues(t *testing.T) {
+	common.TempSuppressOnMemStorage = true
+
 	// clear all state of DB
-	os.Remove("/tmp/todo.db")
-	os.Remove("/tmp/todo.log")
+	if !common.EnableOnMemStorage || common.TempSuppressOnMemStorage == true {
+		os.Remove("/tmp/todo.db")
+		os.Remove("/tmp/todo.log")
+	}
 
 	db := samehada.NewSamehadaDB("/tmp/todo", 200)
 	db.ExecuteSQL("CREATE TABLE name_age_list(name VARCHAR(256), age INT);")
@@ -205,4 +230,6 @@ func TestRebootAndReturnIFValues(t *testing.T) {
 	testingpkg.SimpleAssert(t, len(results3) == 5)
 
 	db3.Shutdown()
+
+	common.TempSuppressOnMemStorage = false
 }
